+++ hawc/services/utils/rasterize.py
